# application/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mechanic API
  description: |
    A comprehensive REST API for managing auto repair shop operations including mechanics, customers, inventory, and service tickets.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. After logging in, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Customer endpoints: 100 requests per hour
    - Login endpoints: 10 requests per minute
    - Specific endpoints may have additional rate limits
    
    ## Error Handling
    The API uses standard HTTP status codes and returns error messages in JSON format.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mechanicapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.mechanic.com/v1
    description: Production server
...
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: Customer login
      description: Authenticate customer and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: customer@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get customer's service tickets
      description: Retrieve all service tickets for the authenticated customer
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Customer tickets retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceTicket'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/:
    get:
      tags:
        - Customers
      summary: Get all customers with pagination
      description: Retrieve paginated list of all customers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            example: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  total:
                    type: integer
                    example: 150
                  pages:
                    type: integer
                    example: 15
                  current_page:
                    type: integer
                    example: 1
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegistration'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/profile:
    get:
      tags:
        - Customers
      summary: Get current customer profile
      description: Retrieve the profile of the authenticated customer
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Customer profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get specific customer by ID
      description: Retrieve a specific customer by their ID
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Customers
      summary: Update customer profile
      description: Update a specific customer by their ID (requires ownership)
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Cannot update other customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Customers
      summary: Delete customer account
      description: Delete a specific customer by their ID (requires ownership)
      security:
        - BearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Customer deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Cannot delete other customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Inventory Routes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Registration failed"
        message:
          type: string
          example: "Internal server error occurred"
        details:
          type: string
          example: "Database connection error"
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Not a valid email address"]
            password: ["Password must be at least 8 characters"]

    # Authentication Schemas
...
    MechanicResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        specialization:
          type: string
          example: "Engine Repair"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
        vehicle_info:
          type: string
          example: "2020 Toyota Camry, VIN: 123456789"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T14:45:00Z"
    CustomerRegistration:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          example: "securepassword123"
        phone:
          type: string
          example: "123-456-7890"
        address:
          type: string
          example: "123 Main St, City, State"
        vehicle_info:
          type: string
          example: "2020 Toyota Camry, VIN: 123456789"
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          example: "John Smith"
        email:
          type: string
          format: email
          example: "john.smith@example.com"
        phone:
          type: string
          example: "987-654-3210"
        address:
          type: string
          example: "456 Oak Ave, City, State"
        vehicle_info:
          type: string
          example: "2021 Honda Civic, VIN: 987654321"
        password:
          type: string
          example: "newpassword123"

    # Inventory Schemas
...
    ServiceTicket:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer_id:
          type: integer
          example: 1
        description:
          type: string
          example: "Engine overheating and strange noises"
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
          example: "pending"
        priority:
          type: string
          enum: [low, normal, high, urgent]
          example: "normal"
        vehicle_info:
          type: string
          example: "2020 Toyota Camry, VIN: 123456789"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T14:45:00Z"
        mechanics:
          type: array
          items:
            $ref: '#/components/schemas/MechanicResponse'
        parts:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: integer
                example: 1
              name:
                type: string
                example: "Engine Oil"
              quantity:
                type: integer
                example: 2
...